fix(deps): add openssl vendored feature
see https://docs.rs/openssl/latest/openssl/#vendored
the issue is when a package was compiled on different
Linux distributions(Arch, Ubuntu, Debian) openssl
crate was depend on the OpenSSL installation directory
which specific to each distribution.

if a rust package was compiled on Ubuntu with open_ssl
crate without `vendored` feature it will not compile
in different distributions(Arch, for example).
it'll throw a error `Failed to find OpenSSL development headers`